Little modification on login display, each editText have an error alert and the email address must respect the format. Fire base adding to the projet for the authentification part
