+++ app/src/main/java/Controler/Authentification.java
@@ -1,0 +1,23 @@
+package Controler;
+
+import android.content.Context;
+
+import com.example.leguidebienpnard.Model.LogActivity;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.InputStream;
+
+public class Authentification {
+
+    private Context context;
+
+    public Authentification(Context context) throws JSONException {
+        this.context = context;
+    }
+
+}
