+++ build.gradle
@@ -9,6 +9,7 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.6.1'
+        classpath 'com.google.gms:google-services:4.3.3'
         
 
         // NOTE: Do not place your application dependencies here; they belong
